+++ components/detail/ladefuchsButton.tsx
